requirements.txt
fastapi
uvicorn[standard]
asyncpg
redis
httpx
sqlparse
pydantic
python-dotenv
pandas
google-genai